spotify-playlist-splitter/requirements.txt
requests
boto3
Pillow